Fix typos and namespaces
