+++ LYBT.Module.Logs/Mapping/LogMappingProfile.cs
