+++ README.md
@@ -1,1 +1,3 @@
 # LYBTZYZS
+
+This repository contains various modules for the LYBT healthcare management system. It is organized as a multi-project .NET solution.
